+++ modules/editors/default.nix
@@ -1,0 +1,9 @@
+{ ... }:
+
+{
+  imports = [
+    ./neovim.nix
+    ./helix.nix
+    ./vscode-server.nix
+  ];
+}