+++ modules/editors/helix.nix
@@ -1,0 +1,8 @@
+{ config, pkgs, userState, ... }:
+
+let
+  enabled = userState.editors.helix or false;
+in
+{
+  environment.systemPackages = pkgs.lib.optional enabled pkgs.helix;
+}